fs: Protect all operations with a Big Honking Lock
FUSE requests are served in goroutines with no concurrency control, so
they can be served concurrently, even for the same Node. To this day,
I am unable to trigger the race detector with this, but fundamentally
we know a lock is needed.

For now, use a single lock in the filesystem. This can later be spread
out to directories and files having their own locks, if and when this
becomes a bottleneck.