fs/clock: Vector Time Pair implementation, no file deletes or dirs yet
Uses a global event counter that's passed in. The Tra sync algorithm
needs file creations to have a time that is greater than any previous
states for a file at the same pathname -- even after the removal of
tombstones and long stretches of time. Tra does this by using a global
clock that ticks once at the beginning of every sync operation. Bazil
will keep a clock in the volume state that advances before every sync.