cas/chunks/chunkutil: Use cgo blake2b for a 5.5x speedup in raw hashing
This translates to about 1.6x improvement in a sequential write
workload currently, but tests show an easy linear scaling for
workloads spread across multiple files can be reached by removing
a locking bottleneck.

With Go 1.2, the pure-Go version faster, and the raw gain was only
1.5x, not as interesting. Compiler changes have ruined the
optimizations in the pure-Go version.