server/http, server, cli/server/run: Serve HTTPS for peers and eventually UI
This commit doesn't yet actually serve the connections.

Call Listen early so we know the port it's listening on.

Explicitly use ListenTCP because we know we'll want to query the port.

The TLS certificate is ephemeral; it will be tied to a long-term
trusted ed25519 key with a custom certificate mechanism, as TLS's PKI
aspects don't seem to bend easily to small copy-pasteable public keys
that would be useful outside of TLS.