fs: Move mutex outside of cleanEpoch
This is needed to avoid races; the caller will want to start a
consistent snapshot view after the cleanEpoch call, but before anyone
uses the new epoch.

To demonstrate the race we need to avoid, assume we have two
goroutines `G1` and `G2` running on a peer, and another peer `P`
synchronizing changes from the first peer:

    G1: cleanEpoch E1->E2
    G2: write file contents F1 at E2
    G1: start view, see F1 at E2
    P: receive sync, merge F1 at E2
    G2: write file contents F2 at E2
    G1: cleanEpoch E2->E3
    G1: start view, see F2 at E2
    P: receive sync, already have file at E2

`P` now thinks it's up to date, but did is not aware of the changes in
`F2`.